requirements.txt
airtable-python-wrapper==0.13.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
importlib-metadata==1.6.0
pyrsistent==0.16.0
requests==2.23.0
six==1.14.0
urllib3==1.25.9
zipp==3.1.0
